requirements.txt
anykeystore==0.2
apex==0.9.10.dev0
certifi==2025.6.15
charset-normalizer==3.4.2
cryptacular==1.6.2
defusedxml==0.7.1
greenlet==3.2.3
hupper==1.12.1
idna==3.10
MarkupSafe==3.0.2
numpy==1.26.4
oauthlib==3.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==25.0
PasteDeploy==3.1.0
pbkdf2==1.3
pillow==11.2.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
pyramid==2.0.2
pyramid-mailer==0.15.1
python3-openid==3.2.0
PyYAML==6.0.1
repoze.sendmail==4.4.1
requests==2.32.4
requests-oauthlib==2.0.0
SQLAlchemy==2.0.41
timm==0.4.12
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.67.1
transaction==5.0
translationstring==1.4
typing_extensions==4.14.0
urllib3==2.4.0
velruse==1.1.1
venusian==3.1.1
WebOb==1.8.9
WTForms==3.2.1
wtforms-recaptcha==0.3.2
zope.deprecation==5.1
zope.interface==7.2
zope.sqlalchemy==3.1
